+++ packages/api/src/settings/types.ts
@@ -9,13 +9,15 @@ export enum SettingGroup {
 
 interface Setting {
   key: string;
+  label: string;
+  description?: string;
+  defaultValue?: string;
   value?: string;
   valueRealized?: string | boolean | number | number[] | string[];
   type: string;
-  label: string;
-  description?: string;
-  group: SettingGroup;
   options?: string;
+  group: SettingGroup;
+  subgroup: number;
 }
 
 export interface SettingDocument extends Setting, Document {}
@@ -23,13 +25,15 @@ export interface SettingDocument extends Setting, Document {}
 export interface SettingDto {
   _id: string;
   key: string;
+  label: string;
+  description?: string;
+  defaultValue?: string;
   value?: string;
   valueRealized?: string | boolean | number | number[] | string[] | object;
   type: string;
-  label: string;
-  description?: string;
-  group: SettingGroup;
   options?: string;
+  group: SettingGroup;
+  subgroup: number;
 }
 
 export interface SettingSetInput {