+++ packages/discord-bot/src/handlers/help.ts
@@ -1,16 +1,20 @@
-import { Collection, CommandInteraction, MessageEmbed } from 'discord.js';
+import {
+  Collection,
+  ChatInputCommandInteraction,
+  EmbedBuilder,
+} from 'discord.js';
 import { Command } from 'src/interfaces/Command';
 
 /**
  * Executes command /help
  *  Provides documentation on how to use the praise discord bot
  *
- * @param {CommandInteraction} interaction
+ * @param {ChatInputCommandInteraction} interaction
  * @param {Collection<string, Command>} commands
  * @returns {Promise<void>}
  */
 export const helpHandler = async (
-  interaction: CommandInteraction,
+  interaction: ChatInputCommandInteraction,
   commands: Collection<string, Command>
 ): Promise<void> => {
   const cmd = interaction.options.getString('command');
@@ -20,12 +24,12 @@ export const helpHandler = async (
       .join('\n');
     await interaction.editReply({
       embeds: [
-        new MessageEmbed()
+        new EmbedBuilder()
           .setTitle('help')
           .setDescription(
             'Welcome to the praise bot!\nTo start using praise, you need to activate praise by using the `/activate` command.\nTo praise, use the `/praise command`.'
           )
-          .addField('Commands', cmdDescription),
+          .addFields({ name: 'Commands', value: cmdDescription }),
       ],
     });
     return;
@@ -34,20 +38,23 @@ export const helpHandler = async (
   if (!commandHelp || !commandHelp.text) {
     await interaction.editReply({
       embeds: [
-        new MessageEmbed()
+        new EmbedBuilder()
           .setTitle(`\`${cmd}\` command`)
           .setDescription('No HelpText found for this command...'),
       ],
     });
     return;
   }
-  const helpEmbed = new MessageEmbed()
+  const helpEmbed = new EmbedBuilder()
     .setTitle(`\`${cmd}\` command`)
     .setDescription(commandHelp.text);
 
   if (commandHelp.subCommands) {
     for (const subCommandHelp of commandHelp.subCommands) {
-      helpEmbed.addField(`${cmd} ${subCommandHelp.name}`, subCommandHelp.text);
+      helpEmbed.addFields({
+        name: `${cmd} ${subCommandHelp.name}`,
+        value: subCommandHelp.text,
+      });
     }
   }