+++ packages/discord-bot/src/utils/buttons/confirmationButtons.ts
@@ -1,11 +1,11 @@
-import { MessageButton } from 'discord.js';
+import { ButtonBuilder, ButtonStyle } from 'discord.js';
 
-export const continueButton = new MessageButton()
+export const continueButton = new ButtonBuilder()
   .setCustomId('continue')
   .setLabel('Continue')
-  .setStyle('SUCCESS');
+  .setStyle(ButtonStyle.Success);
 
-export const cancelButton = new MessageButton()
+export const cancelButton = new ButtonBuilder()
   .setCustomId('cancel')
   .setLabel('Cancel')
-  .setStyle('DANGER');
+  .setStyle(ButtonStyle.Danger);