+++ packages/frontend/src/utils/polyfills.ts
@@ -1,0 +1,12 @@
+/**
+ * This is used for loading Rainbow Kit
+ * Wallet Connect option for logging
+ */
+
+import { Buffer } from 'buffer';
+
+window.global = window.global ?? window;
+window.Buffer = window.Buffer ?? Buffer;
+window.process = window.process ?? { env: {} }; // Minimal process polyfill
+
+export {};